python/requirements.txt
web3>=6.15.1
eth-account>=0.11.3
python-dotenv>=1.0.1
rich>=13.7.1
